logs: split tracing config out from common
