+++ Cargo.toml
@@ -15,6 +15,7 @@ circus-admin        = { path = "./crates/admin-cli", version = "0.11.0" }
 circus-agent        = { path = "./crates/agent", version = "0.11.0" }
 circus-common       = { path = "./crates/common", version = "0.11.0" }
 circus-evaluator    = { path = "./crates/evaluator", version = "0.11.0" }
+circus-logs         = { path = "./crates/logs", version = "0.11.0" }
 circus-migrations   = { path = "./crates/migrations", version = "0.11.0" }
 circus-proto        = { path = "./crates/proto", version = "0.11.0" }
 circus-queue-runner = { path = "./crates/queue-runner", version = "0.11.0" }
