+++ crates/common/Cargo.toml
@@ -6,37 +6,37 @@ license.workspace    = true
 repository.workspace = true
 
 [dependencies]
-argon2.workspace             = true
-chrono.workspace             = true
-circus-migrations.workspace  = true
-clap.workspace               = true
-color-eyre.workspace         = true
-config.workspace             = true
-git2.workspace               = true
-hex.workspace                = true
-hmac.workspace               = true
-humantime-serde.workspace    = true
-lettre.workspace             = true
-libc.workspace               = true
-password-hash.workspace      = true
-percent-encoding.workspace   = true
-regex.workspace              = true
-reqwest.workspace            = true
-rustls.workspace             = true
-serde.workspace              = true
-serde_json.workspace         = true
-sha2.workspace               = true
-sqlx.workspace               = true
-tempfile.workspace           = true
-thiserror.workspace          = true
-tokio.workspace              = true
-toml.workspace               = true
-tracing.workspace            = true
-tracing-subscriber.workspace = true
-url.workspace                = true
-urlencoding.workspace        = true
-uuid.workspace               = true
-whoami.workspace             = true
+argon2.workspace            = true
+chrono.workspace            = true
+circus-logs.workspace       = true
+circus-migrations.workspace = true
+clap.workspace              = true
+color-eyre.workspace        = true
+config.workspace            = true
+git2.workspace              = true
+hex.workspace               = true
+hmac.workspace              = true
+humantime-serde.workspace   = true
+lettre.workspace            = true
+libc.workspace              = true
+password-hash.workspace     = true
+percent-encoding.workspace  = true
+regex.workspace             = true
+reqwest.workspace           = true
+rustls.workspace            = true
+serde.workspace             = true
+serde_json.workspace        = true
+sha2.workspace              = true
+sqlx.workspace              = true
+tempfile.workspace          = true
+thiserror.workspace         = true
+tokio.workspace             = true
+toml.workspace              = true
+tracing.workspace           = true
+url.workspace               = true
+urlencoding.workspace       = true
+uuid.workspace              = true
+whoami.workspace            = true
 
 [lints]
 workspace = true
