+++ crates/logs/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name                 = "circus-logs"
+version.workspace    = true
+edition.workspace    = true
+license.workspace    = true
+repository.workspace = true
+
+[dependencies]
+serde.workspace              = true
+tracing-subscriber.workspace = true
+
+[lints]
+workspace = true
