+++ crates/migrate-cli/Cargo.toml
@@ -11,11 +11,11 @@ path = "src/main.rs"
 
 [dependencies]
 circus-common.workspace = true
+circus-logs.workspace   = true
 
-clap.workspace               = true
-color-eyre.workspace         = true
-tokio.workspace              = true
-tracing-subscriber.workspace = true
+clap.workspace       = true
+color-eyre.workspace = true
+tokio.workspace      = true
 
 [dev-dependencies]
 tempfile.workspace = true
